Add an option to telemetry_utils to not delete parent directory of chartjson file.
The android test runner automatically cleans up the parent directory and moves the file to /tmp, so on android this code was deleting /tmp

BUG=441446

Review URL: https://codereview.chromium.org/797123002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@293376 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/telemetry_utils.py
@@ -20,9 +20,10 @@ from slave.performance_log_processor import _FormatHumanReadable
 
 class TelemetryResultsProcessor(object):
 
-  def __init__(self, filename, is_ref):
+  def __init__(self, filename, is_ref, cleanup_parent_dir):
     self._chart_filename = filename
     self._is_reference_build = is_ref
+    self._cleanup_parent_dir = cleanup_parent_dir
 
   def ChartJson(self):
     try:
@@ -38,11 +39,12 @@ class TelemetryResultsProcessor(object):
     except OSError:
       logging.error('Unable to remove telemetry output file %s',
                     self._chart_filename)
-    try:
-      os.rmdir(os.path.dirname(self._chart_filename))
-    except OSError:
-      logging.error('Unable to remove telemetry output dir %s',
-                    os.path.dirname(self._chart_filename))
+    if self._cleanup_parent_dir:
+      try:
+        os.rmdir(os.path.dirname(self._chart_filename))
+      except OSError:
+        logging.error('Unable to remove telemetry output dir %s',
+                      os.path.dirname(self._chart_filename))
 
   def IsChartJson(self):
     """This is the new telemetry --chartjson output format."""
